Migrate and test InvalidHttpResponseException
